move 404 to shared ui
