--- main-site/src/pageStyles/globals.css
@@ -1,36 +1,0 @@
-.day-background-desktop {
-  background-image: url(../../../shared-ui/images/day-background-desktop.svg);
-  background-size: cover;
-}
-
-.dark-background-desktop {
-  background-image: url(../../../shared-ui/images/dark-background-desktop.svg);
-  background-size: cover;
-}
-
-.mobile-light-background {
-  background-image: url(../../../shared-ui/images/mobile-light-background.svg);
-  background-size: cover;
-}
-
-.mobile-dark-background {
-  background-image: url(../../../shared-ui/images/mobile-dark-background.svg);
-  background-size: cover;
-}
-
-.error-background-desktop {
-  background-image: url(../../../shared-ui/images/error-desktop-background.svg);
-  background-size: cover;
-  height: 100vh;
-}
-
-.error-background-mobile {
-  background-image: url(../../../shared-ui/images/error-mobile-background.svg);
-  background-size: cover;
-  height: 100vh;
-}
-
-html,
-body {
-  max-width: 1920px;
-}
